agents/requirements.txt
# Web3 / Blockchain (Flare Coston2)
web3>=7.0.0
eth-account>=0.13.0

# Core (All scripts)
python-dotenv>=1.0.0
pydantic>=2.9.0

# Server (Butler API)
uvicorn>=0.32.0
fastapi>=0.115.0

# AI / Vector DB (Butler + Seeding)
openai>=1.50.0
qdrant-client>=1.7.0
mem0ai>=0.0.10

# LangGraph (Butler Agent Orchestration)
langgraph>=0.2.0
langchain-core>=0.3.0

# HTTP tools
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.10.0

# Database (PostgreSQL)
asyncpg>=0.29.0

# Scraping (Hackathon Agent)
beautifulsoup4>=4.12.0

# Browser Automation (Hackathon Registration)
playwright>=1.40.0

# Document Parsing (CV Magic Agent)
pypdf>=4.0.0
python-docx>=1.1.0

# Optional: Agent-specific tools
# twilio>=9.3.0  # For Caller Agent

# Dev
pytest>=8.0.0
pytest-asyncio>=0.24.0
